Fix getPersonByEmail
This method is a legacy (from fling) "helper" wrapper to lookup a user by email via the /people API.    Some of these old helpers were not added to the tests and this is one of them.  It looks like JP attempted some cleanup and introduced an undeclared variable.  This fix should address this.